Right-click cursor tp fix & Lock Border Improvement
With this Pull Request, I am implementing a third feature to fix the teleportation of the mouse cursor upon holding the right-click key and dragging. It might not work 100% the same like the official roblox client, but I tried my best replicating it. That being said, it might not work 100% of the time, but you can also test to see wether it works or not

I am also improving the Lock Border feature to keep and teleport the mouse further to the border of the window instead of ~20-30 pixels inside the window allowing for more room, however the top is a complicated area which I improved a bit, but mostly left out the way it was because I realized upon testing that it can get stuck at the very top and this can be especially annoying in fullscreen mode. 

With the changed values for Lock Border, I also took the same logic from there and added it to my third feature to confine the mouse cursor as well while fixing the teleportation issue when right-clicking and dragging the avatars' camera.